+++ models/layers/feedforwards/leff.py
@@ -12,7 +12,7 @@ from einops import rearrange
 
 
 class LeFFBlock(nn.Module):
-    expand_ratio: int = None
+    expand_ratio: float = None
     hidden_ch: int = None
     kernel_size: int = 5
     activation_fn: Callable = nn.activation.gelu
@@ -34,7 +34,7 @@ class LeFFBlock(nn.Module):
                     'Must provide one of expand_ratio or hidden_ch')
             hidden_ch = self.hidden_ch
         else:
-            hidden_ch = max(1, self.expand_ratio * in_ch)
+            hidden_ch = max(1, int(self.expand_ratio * in_ch))
 
         dense = partial(nn.Dense,
                         use_bias=True,
